Refactor: 파일 정리 및 이미지 절대 경로로 변경
1. 파일이름 수정(명시적으로 표시)
2. 이미지 경로 수정(상대경로->절대경로)
3. 안쓰는 코드 삭제
