+++ src/components/shopping-section/mainPage/ResentSlideList.tsx
@@ -1,27 +1,19 @@
+/* eslint-disable react/jsx-pascal-case */
 import React, { useEffect, useState } from 'react';
 import { Swiper, SwiperSlide } from 'swiper/react';
 import axios from 'axios';
-import eyes from '../assets/eyes.png';
-import SwiperCore, {
-    Navigation,
-    Pagination,
-    Scrollbar,
-    A11y,
-} from 'swiper/modules';
+import eyes from 'assets/eyes.png';
+import { Navigation, Pagination } from 'swiper/modules';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { faHeart } from '@fortawesome/free-regular-svg-icons';
 import 'swiper/swiper-bundle.css';
 import styled from 'styled-components';
-import shopping1 from '../assets/shopping1.jpg';
-import shopping2 from '../assets/shopping2.jpg';
-import shopping3 from '../assets/shopping3.jpg';
-import shopping4 from '../assets/shopping4.jpg';
-import shopping5 from '../assets/shopping5.jpg';
+
 import {
     useBuyModalStore,
     useModalStore,
     useProductId,
-} from '../store/shoppingHeaderSelectBarStore';
+} from '../../../store/shoppingHeaderSelectBarStore';
+import { ShoppingDetailHeader } from '../detailPage/ShoppingDetailModal';
 
 interface ApiResponse {
     discountedPrice: number;
@@ -67,8 +59,8 @@ export default function ShoppingSlideResent() {
     useEffect(() => {
         const getResentProduct = async () => {
             try {
-                const res = await axios.get<ApiResponse, any>(
-                    `${process.env.REACT_APP_API_KEY}/shop/best`,
+                const res = await axios.get<ApiResponse[]>(
+                    `${process.env.REACT_APP_API_KEY}/shop/new`,
                 );
 
                 setResentProductList(res.data);
@@ -78,6 +70,7 @@ export default function ShoppingSlideResent() {
         };
         getResentProduct();
     }, []);
+    console.log(resentProductList);
 
     const onClickToggleModal = (
         changedCategory: string,
@@ -115,7 +108,7 @@ export default function ShoppingSlideResent() {
                             slidesPerGroup: 4,
                             //   spaceBetween: 10,
                         },
-                        // 767px 이상일 때
+                        // 768px 이상일 때
                         767: {
                             slidesPerView: 3,
                             spaceBetween: 50,
@@ -186,6 +179,29 @@ export default function ShoppingSlideResent() {
                     })}
                 </_customSwiper>
             </_swiperWrapper>
+            {isOpenModal && (
+                <ShoppingDetailHeader
+                    onClickToggleModal={(title, category, productId) =>
+                        console.log(title, category, productId)
+                    }
+                    category={category}
+                    title={title}
+                    productId={productId}
+                ></ShoppingDetailHeader>
+            )}
+
+            {isOpenModal && (
+                <_backdrop
+                    onClick={(e: React.MouseEvent) => {
+                        e.preventDefault();
+                        setOpenModal(false);
+                        setIsOpenBuyModal(false);
+                        if (onClickToggleModal) {
+                            onClickToggleModal(category, title, productId);
+                        }
+                    }}
+                ></_backdrop>
+            )}
         </>
     );
 }
@@ -382,4 +398,13 @@ const _realPriceDiv = styled.div`
     @media (max-width: 767px) {
         font-size: 15px;
     }
+`;
+
+const _backdrop = styled.div`
+    width: 100vw;
+    height: 100vh;
+    position: fixed;
+    top: 0;
+    z-index: 3;
+    background-color: rgba(0, 0, 0, 0.2);
 `;
